Refactor repository imports and enhance MockCollection functionality
